+++ onmt/train/Optim.lua
@@ -61,12 +61,17 @@ local options = {
   {'-learning_rate',       1   , [[Starting learning rate. If adagrad or adam is used,
                                       then this is the global learning rate. Recommended settings are: sgd = 1,
                                       adagrad = 0.1, adam = 0.0002]]},
+  {'-min_learning_rate',   0   , [[Do not continue the training past this learning rate]]},
   {'-max_grad_norm',       5   , [[If the norm of the gradient vector exceeds this renormalize it to have
                                        the norm equal to max_grad_norm]]},
-  {'-learning_rate_decay', 0.5 , [[Decay learning rate by this much if (i) perplexity does not decrease
-                                       on the validation set or (ii) epoch has gone past the start_decay_at_limit]]},
-  {'-start_decay_at',      9   , [[Start decay after this epoch]],
-                                 {valid=onmt.utils.ExtendedCmdLine.isUInt()}}
+  {'-learning_rate_decay', 0.5 , [[Learning rate decay factor]]},
+  {'-start_decay_at',      9   , [[With 'default' decay mode, start decay after this epoch]],
+                                 {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
+  {'-start_decay_ppl_delta', 0 , [[Start decay when validation perplexity improvement is lower than this value]]},
+  {'-decay',          'default', [[When to apply learning rate decay.
+                                 'default': decay after each epoch past start_decay_at or as soon as the validation perplexity is not improving more than start_decay_ppl_delta,
+                                 'perplexity_only': only decay when validation perplexity is not improving more than start_decay_ppl_delta]],
+                                 {enum={'default', 'perplexity_only'}}}
 }
 
 function Optim.declareOpts(cmd)
@@ -139,6 +144,10 @@ end
 
 -- decay learning rate if val perf does not improve or we hit the startDecayAt limit
 function Optim:updateLearningRate(score, epoch)
+  local function decayLr()
+    self.args.learning_rate = self.args.learning_rate * self.args.learning_rate_decay
+  end
+
   if self.args.optim == 'sgd' then
     self.valPerf[#self.valPerf + 1] = score
 
@@ -146,18 +155,27 @@ function Optim:updateLearningRate(score, epoch)
       self.startDecay = true
     end
 
+    local decayConditionMet = false
+
     if self.valPerf[#self.valPerf] ~= nil and self.valPerf[#self.valPerf-1] ~= nil then
       local currPpl = self.valPerf[#self.valPerf]
       local prevPpl = self.valPerf[#self.valPerf-1]
-      if currPpl > prevPpl then
+      if prevPpl - currPpl < self.args.start_decay_ppl_delta then
         self.startDecay = true
+        decayConditionMet = true
       end
     end
 
-    if self.startDecay then
-      self.args.learning_rate = self.args.learning_rate * self.args.learning_rate_decay
+    if self.args.decay == 'default' and self.startDecay then
+      decayLr()
+    elseif self.args.decay == 'perplexity_only' and decayConditionMet then
+      decayLr()
     end
+
+    return self.args.learning_rate >= self.args.min_learning_rate
   end
+
+  return true
 end
 
 function Optim:getLearningRate()